Add possibility to flush output immediately
Before output was only handled after each expression were finished
executing (or a message or plot was created). This meant that if a
longrunning loop was executed (loop = only one expression), the
output wasn't handled until it was finished. In irkernel, which
sends output over the wire when it becomes available, this meant
that the UI wasn't getting any updates.

Now, if `evaluate(...., enable_flush = TRUE)`, you can insert a
`flush(stdout())` or a `flush.console()` to flush the output and
make text output available to the handler.