Use restarts for control flow (#171)
Restarts are a (very) rarely used technique that allows for non-local control flow, i.e. the ability to jump immediately to a different destination in a calling function. This allows us to direct what should happen in the error calling handler (which usually can't affect control flow) based on the value of `on_error`.