Better function parsing
The original intent of this code (as much as I can remember) was to make it easier to generate multiline examples by wrapping them in an anonymous zero-argument function (the simplest structure that preserves srcrefs). But I'm not sure the code ever worked properly so I reimplemented it from first principles a considerably cleaner approach.

This is a robust alternative to the testing helper `evaluate_()`, allowing me to eliminate that helper.